# Rooted*: make public APIs inline