By default, Rust doesn't inline calls across crate boundaries. Since these functions are small, called frequently, and from across a crate boundary, we probably want them to be inline. We force this by marking them `#[inline]`, which is also done in the analogous functions in `std`.

A quick local benchmark using the tor minimal test shows there might be a measurable improvement here, though it's inconclusive since the confidence intervals overlap substantially:

Before:

```
Benchmark 1: ./setup test --extra tor-minimal
  Time (mean ± σ):     17.424 s ±  0.455 s    [User: 18.446 s, System: 6.024 s]
  Range (min … max):   16.818 s … 18.123 s    10 runs
```

After:

```
Benchmark 1: ./setup test --extra tor-minimal
  Time (mean ± σ):     17.119 s ±  0.482 s    [User: 18.211 s, System: 5.993 s]
  Range (min … max):   16.534 s … 18.020 s    10 runs
```