@sporksmith 2023-02-14
Started a larger tor benchmark, mostly out of curiosity. I'm fairly confident this should be performance-neutral at worst. https://github.com/shadow/benchmark/actions/runs/4176543974